+++ tests/test_arche.py
@@ -248,9 +248,7 @@ def test_data_quality_report(mocker):
 
 def test_compare_with_customized_rules(mocker, get_job_items):
     mocked_save_result = mocker.patch("arche.Arche.save_result", autospec=True)
-    mocked_coverage = mocker.patch(
-        "arche.rules.category_coverage.compare_coverage_per_category", autospec=True
-    )
+    mocked_coverage = mocker.patch("arche.rules.category.get_difference", autospec=True)
     mocked_price_url = mocker.patch(
         "arche.rules.price.compare_prices_for_same_urls", autospec=True
     )
@@ -267,7 +265,7 @@ def test_compare_with_customized_rules(mocker, get_job_items):
     arche.compare_with_customized_rules(source_items, target_items, {})
 
     mocked_coverage.assert_called_once_with(
-        source_items.key, target_items.key, source_items.df, target_items.df, {}
+        source_items.key, target_items.key, source_items.df, target_items.df, []
     )
     mocked_price_url.assert_called_once_with(source_items.df, target_items.df, {})
     mocked_name_url.assert_called_once_with(source_items.df, target_items.df, {})
@@ -277,9 +275,7 @@ def test_compare_with_customized_rules(mocker, get_job_items):
 
 
 def test_compare_with_customized_rules_none_target(mocker, get_job_items):
-    mocked_coverage = mocker.patch(
-        "arche.rules.category_coverage.compare_coverage_per_category", autospec=True
-    )
+    mocked_coverage = mocker.patch("arche.rules.category.get_difference", autospec=True)
     arche = Arche("key")
     assert not arche.compare_with_customized_rules(
         source_items=get_job_items, target_items=None, tagged_fields={}